Add new option `--secret` to `get-oauth2-token.py`
This commit adds an option to obtain refresh token from "secret service" (GNOME Keyring,
KeepassXC and perhaps others). This is more convenient as it allows multiple
applications (email client, git-send-email) to use the same refresh token and
maybe more secure as the token is stored in app designed to store secrets rather
than in files on the disk.

For now however, it only works with refresh tokens stored by Evolution.